Refactor PublishPublicLibrariesPlugin
- Extract GeneratePom class into separate file
