+++ buildSrc/src/main/kotlin/org/gradle/plugins/publish/PublishPublicLibrariesPlugin.kt
@@ -16,23 +16,19 @@
 package org.gradle.plugins.publish
 
 import accessors.base
-import accessors.java
 import accessors.groovy
+import accessors.java
 import groovy.lang.MissingPropertyException
-import org.gradle.api.DefaultTask
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.artifacts.ExternalDependency
-import org.gradle.api.artifacts.ProjectDependency
-import org.gradle.api.artifacts.maven.Conf2ScopeMappingContainer
 import org.gradle.api.internal.artifacts.publish.DefaultPublishArtifact
-import org.gradle.api.plugins.MavenRepositoryHandlerConvention
-import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.Upload
 import org.gradle.api.tasks.bundling.Jar
+import org.gradle.kotlin.dsl.creating
+import org.gradle.kotlin.dsl.extra
+import org.gradle.kotlin.dsl.get
+import org.gradle.kotlin.dsl.getting
 import org.gradle.kotlin.dsl.*
-import java.io.File
 import java.util.*
 
 open class PublishPublicLibrariesPlugin : Plugin<Project> {
@@ -49,6 +45,22 @@ open class PublishPublicLibrariesPlugin : Plugin<Project> {
             from(main.java.srcDirs + main.groovy.srcDirs)
         }
 
+        configureUploadArchivesTask(generatePom)
+
+        artifacts {
+            add(generatePom.publishRuntime.name, tasks["jar"])
+            add(generatePom.publishRuntime.name, sourceJar)
+            add(generatePom.publishRuntime.name, DefaultPublishArtifact(base.archivesBaseName,
+                "pom",
+                "pom",
+                null,
+                Date(),
+                generatePom.pomFile,
+                generatePom))
+        }
+    }
+
+    private fun Project.configureUploadArchivesTask(generatePom: GeneratePom) {
         val uploadArchives by tasks.getting(Upload::class) {
             // TODO Add magic property to upcoming configuration interface
             onlyIf { !project.hasProperty("noUpload") }
@@ -69,19 +81,6 @@ open class PublishPublicLibrariesPlugin : Plugin<Project> {
                 }
             }
         }
-
-        artifacts {
-            add(generatePom.publishRuntime.name, tasks["jar"])
-            add(generatePom.publishRuntime.name, sourceJar)
-            add(generatePom.publishRuntime.name, DefaultPublishArtifact(base.archivesBaseName,
-                "pom",
-                "pom",
-                null,
-                Date(),
-                generatePom.pomFile,
-                generatePom))
-        }
-
     }
 
     private
@@ -96,7 +95,8 @@ open class PublishPublicLibrariesPlugin : Plugin<Project> {
     val Project.groupId: String
         get() = group.toString().replace("\\.", "/")
 
-    private fun Project.failEarlyIfCredentialsAreNotSet(upload: Upload) {
+    private
+    fun Project.failEarlyIfCredentialsAreNotSet(upload: Upload) {
         gradle.taskGraph.whenReady({
             if (hasTask(upload)) {
                 if (artifactoryUserName.isNullOrEmpty()) {
@@ -108,7 +108,6 @@ open class PublishPublicLibrariesPlugin : Plugin<Project> {
             }
         })
     }
-
 
     // TODO Add magic property to upcoming configuration interface
     private
@@ -125,56 +124,6 @@ open class PublishPublicLibrariesPlugin : Plugin<Project> {
         get() = java.sourceSets["main"]
 }
 
-open class GeneratePom : DefaultTask() {
-    @InputFile
-    val pomFile = File(temporaryDir, "pom.xml")
-
-    // TODO How to annotate?
-    val publishCompile by project.configurations.creating
-
-    // TODO How to annotate?
-    val publishRuntime by project.configurations.creating
-
-    init {
-        // Subprojects assign dependencies to publishCompile to indicate that they should be part of the published pom.
-        // Therefore compile needs to contain those dependencies and extend publishCompile
-        val compile by project.configurations.getting {
-            extendsFrom(publishCompile)
-        }
-    }
-
-    @TaskAction
-    fun generatePom(): Unit = project.run {
-        addDependenciesToPublishConfigurations()
-        val install by tasks.getting(Upload::class)
-        install.repositories {
-            withConvention(MavenRepositoryHandlerConvention::class) {
-                mavenInstaller {
-                    pom {
-                        scopeMappings.mappings.clear()
-                        scopeMappings.addMapping(300, publishRuntime, Conf2ScopeMappingContainer.RUNTIME)
-                        groupId = project.group.toString()
-                        artifactId = base.archivesBaseName
-                        version = project.version.toString()
-                        writeTo(pomFile)
-                    }
-                }
-            }
-        }
-    }
-
-    private
-    fun Project.addDependenciesToPublishConfigurations() {
-        dependencies {
-            publishCompile.allDependencies.withType<ProjectDependency>().forEach {
-                publishRuntime("org.gradle:${it.dependencyProject.base.archivesBaseName}:$version")
-            }
-            publishCompile.allDependencies.withType<ExternalDependency>().forEach {
-                publishRuntime(it)
-            }
-        }
-    }
-}
